Fix object ordering bug
The implementation of World hit() wasn't quite right. It would return
prematurely as soon as there was a hit, so the order objects were
added to the world mattered. This change fixes that, and allows
objects to be added in any order -- the front-most object will always
be in front.